# swagger/swagger.yml
swagger: "2.0"
info:
  description: ""
  version: "0.0.1"
  title: "prodx"
  contact:
    email: "alexandre@ineosystems.com.br"
host: "127.0.0.1:3000"
basePath: "/v1"
...
schemes:
- "http"
securityDefinitions:
  APIKeyHeader:
    type: apiKey
    in: header
    name: x-api-token
paths:
...
  /keypair:
    post:
      tags:
      - "keypair"
      summary: "Add a new keypair"
      description: ""
      operationId: "addKeypair"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      security:
        - APIKeyHeader: []
      parameters:
      - in: "body"
        name: "body"
        description: "Keypair object that needs to be added"
        required: true
        schema:
          $ref: "#/definitions/Keypair"
      responses:
        201:
          description: "Created"
          schema:
            type: integer
        405:
          description: "Invalid input"
          schema:
            $ref: "#definitions/ApiResponse"
        409:
          description: "Already exists"
          schema:
            $ref: "#definitions/ApiResponse"
        500:
          description: "Internal error"
          schema:
            $ref: "#definitions/ApiResponse"
    put:
      tags:
      - "keypair"
      summary: "Update an existing keypair"
      description: ""
      operationId: "updateKeypair"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      security:
        - APIKeyHeader: []
      parameters:
      - in: "body"
        name: "body"
        description: "Keypair object that needs to be updated"
        required: true
        schema:
          $ref: "#/definitions/Keypair"
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "Keypair not found"
        405:
          description: "Validation exception"
  /keypairs:
    get:
      tags:
      - "keypair"
      summary: "Finds Keypair by customer"
      operationId: "findKeypairByCustomer"
      security:
        - APIKeyHeader: []
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: array
            items:
              $ref: "#/definitions/Keypair"
        400:
          description: "Invalid customer id"
        404:
          description: "Keypair not found"
        500:
          description: "InternalError"
          schema:
            $ref: "#definitions/ApiResponse"
  /keypair/{keypair_id}:
    get:
      tags:
      - "keypair"
      summary: "Find keypair by ID"
      description: "Returns a single keypair"
      operationId: "getKeypairById"
      produces:
      - "application/json"
      security:
        - APIKeyHeader: []
      parameters:
      - name: "keypair_id"
        in: "path"
        description: "ID of keypair to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Keypair"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Keypair not found"
        500:
          description: "InternalError"
          schema:
            $ref: "#definitions/ApiResponse"
    post:
      tags:
      - "keypair"
      summary: "Updates a keypair with form data"
      description: ""
      operationId: "updateKeypairWithForm"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      security:
        - APIKeyHeader: []
      parameters:
      - name: "keypair_id"
        in: "path"
        description: "ID of keypair that needs to be updated"
        required: true
        type: "integer"
        format: "int64"
      - name: "public_key"
        in: "formData"
        description: "Updated name of the keypair"
        required: false
        type: "string"
      responses:
        405:
          description: "Invalid input"
    delete:
      tags:
      - "keypair"
      summary: "Deletes a keypair"
      description: ""
      operationId: "deleteKeypair"
      produces:
      - "application/json"
      security:
        - APIKeyHeader: []
      parameters:
      - name: "api_key"
        in: "header"
        required: false
        type: "string"
      - name: "keypair_id"
        in: "path"
        description: "Keypair id to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "Keypair not found"
...
  /cell/{cell_id}/keypair/{keypair_name}:
    post:
      tags:
      - "keypair"
      summary: "Add a keypair to cell"
      description: ""
      operationId: "addCellKeypair"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      security:
        - APIKeyHeader: []
      parameters:
      - name: "cell_id"
        in: "path"
        description: "ID of cell that will be used"
        required: true
        type: "integer"
        format: "int64"
      - name: "keypair_name"
        in: "path"
        description: "Name of keypair that will be used"
        required: true
        type: "string"
      responses:
        201:
          description: "Created"
          schema:
            type: integer
        404:
          description: "does not exists"
        405:
          description: "Invalid input"
          schema:
            $ref: "#definitions/ApiResponse"
        409:
          description: "Already exists"
          schema:
            $ref: "#definitions/ApiResponse"
        500:
          description: "Internal error"
          schema:
            $ref: "#definitions/ApiResponse"
...
definitions:
...
  Keypair:
    type: "object"
    required:
      - name
      - public_key
    properties:
      id:
        type: integer
      name:
        type: "string"
      public_key:
        type: "string"
...